pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "btc-bybit-alert-engine"
version = "0.5.0"
description = "Bybit-native BTC alert engine with intrabar simulation, manual history import, pilot registries, backfill, walk-forward research, and verification"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0",
  "pydantic>=2.7",
  "websockets>=12.0",
  "httpx>=0.27",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.4",
]

[project.optional-dependencies]
research = [
  "lightgbm>=4.0",
]
dev = [
  "pytest>=8.0",
]
viz = [
  "altair>=5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[project.scripts]
btc-alert-engine = "btc_alert_engine.cli:main"
